feat(wb-handler): add data save feature for samp
Added a feature where the sample data grabbed by the data puller is properly placed and saved in the excel workbook. A logger has also been added. A function has been placed for getting the active sheet name. a sys/exit() was added just before save to temporarily disable overwriting important work during development.

BREAKING CHANGE: The sys.exit() will prevent the data from saving in the workbook. This is on line 378 in WorkbookHandler.py. Please remove or uncomment the line if you intend to change the code to save data. This has been set as a precaution.