fix(loop): fix db and wb handling
fixed how the workbook functions are used and update a parameter to handle sample data. Prints have been changed to logs. Moved database connection update info.

BREAKING CHANGE: with the addition of a new use for the workbook handler and database swapping, there is a need to ensure that proper data is being passed to workbook. Currently some functionality is unpredictable as older versions of the workbook may exist.